Move namespace and add listener
