+++ diagnostic-listener/DiagnosticListenerExample/Program.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 
-namespace DiagnosticListener
+namespace DiagnosticListenerExample
 {
     public class Program
     {
