+++ diagnostic-listener/DiagnosticListenerExample/Properties/launchSettings.json
