+++ diagnostic-listener/DiagnosticListenerExample/project.json
@@ -11,7 +11,9 @@
     "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.0.0",
     "Microsoft.Extensions.Configuration.FileExtensions": "1.0.0",
     "Microsoft.Extensions.Configuration.Json": "1.0.0",
-    "Microsoft.Extensions.Configuration.CommandLine": "1.0.0"
+    "Microsoft.Extensions.Configuration.CommandLine": "1.0.0",
+    "Microsoft.Extensions.DiagnosticAdapter": "1.1.0",
+    "System.Diagnostics.DiagnosticSource": "4.3.0"
   },
 
   "tools": {
